refactor the category page
